Fixed NullReferenceException of bookSearch/API
Fixed NullReferenceException due to bookSearch has use null value to
operate on method in String class which has illegal operation.
To resolve this,we reimplement String.Contains method by create
StringUtil class and reimplement it for use instead of direct
String.Contains()
Note:Some atrribute of book can NULLable string which is cause to this
error so.